# server/openapi.json
{"openapi":"3.1.0","info":{"title":"ALT系统API","description":"这是一个ALT系统的API接口","version":"0.1.0"},"paths":{"/api/llm/":{"get":{"tags":["LLM接口"],"summary":"Llm Ui","operationId":"llm_ui_api_llm__get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/api/llm/chat":{"post":{"tags":["LLM接口"],"summary":"Chat","description":"同步聊天接口","operationId":"chat_api_llm_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/llm/stream-chat":{"post":{"tags":["LLM接口"],"summary":"Stream Chat","description":"流式聊天接口","operationId":"stream_chat_api_llm_stream_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tts/text-to-audio":{"post":{"tags":["TTS接口"],"summary":"Text To Audio","description":"文本转语音API接口\n\n参数:\n- text: 要转换的文本内容\n- filename: 输出音频文件名（可选，默认为output.wav）\n\n返回:\n- 成功: {\"success\": True, \"filename\": \"生成的音频文件名\", \"message\": \"音频生成成功\"}\n- 失败: {\"success\": False, \"message\": \"错误信息\"}","operationId":"text_to_audio_api_tts_text_to_audio_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TextToAudioRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Text To Audio Api Tts Text To Audio Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tts/audio/{filename}":{"get":{"tags":["TTS接口"],"summary":"Get Audio","description":"获取生成的音频文件\n\n参数:\n- filename: 音频文件名\n\n返回:\n- 音频文件","operationId":"get_audio_api_tts_audio__filename__get","parameters":[{"name":"filename","in":"path","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tts/text-to-audio-direct":{"post":{"tags":["TTS接口"],"summary":"Text To Audio Direct","description":"文本转语音并直接返回音频文件\n\n参数:\n- text: 要转换的文本内容\n- filename: 输出音频文件名（可选，默认为output.wav）\n\n返回:\n- 生成的音频文件","operationId":"text_to_audio_direct_api_tts_text_to_audio_direct_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TextToAudioRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tts/platform":{"get":{"tags":["TTS接口"],"summary":"Get Tts Platform","description":"获取当前使用的TTS平台\n\n返回:\n- 当前使用的TTS平台名称","operationId":"get_tts_platform_api_tts_platform_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/asr/audio-to-text":{"post":{"tags":["ASR接口"],"summary":"Audio To Text Api","description":"音频转文本API接口（基于本地文件路径）\n\n参数:\n- audio_path: 音频文件路径\n- model_size: 模型大小，可选: 'tiny', 'base', 'small', 'medium', 'large-v2', 'large-v3'（默认: medium）\n- language: 语言代码，如 'zh' 中文, 'en' 英文，设为 None 可自动检测（默认: None）\n- device: 运行设备，'cuda'（GPU）、'cpu'，'auto' 自动选择（默认: auto）\n\n返回:\n- 成功: {\"success\": True, \"text\": \"识别出的文本\", \"message\": \"音频识别成功\"}\n- 失败: {\"success\": False, \"message\": \"错误信息\"}","operationId":"audio_to_text_api_api_asr_audio_to_text_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AudioToTextRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Audio To Text Api Api Asr Audio To Text Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/asr/upload-and-transcribe":{"post":{"tags":["ASR接口"],"summary":"Upload And Transcribe","description":"上传音频文件并进行语音识别API接口\n\n参数:\n- file: 要上传的音频文件\n- model_size: 模型大小，可选: 'tiny', 'base', 'small', 'medium', 'large-v2', 'large-v3'（默认: medium）\n- language: 语言代码，如 'zh' 中文, 'en' 英文，设为 None 可自动检测（默认: None）\n- device: 运行设备，'cuda'（GPU）、'cpu'，'auto' 自动选择（默认: auto）\n\n返回:\n- 成功: {\"success\": True, \"text\": \"识别出的文本\", \"message\": \"音频识别成功\"}\n- 失败: {\"success\": False, \"message\": \"错误信息\"}","operationId":"upload_and_transcribe_api_asr_upload_and_transcribe_post","parameters":[{"name":"model_size","in":"query","required":false,"schema":{"type":"string","default":"medium","title":"Model Size"}},{"name":"language","in":"query","required":false,"schema":{"type":"string","title":"Language"}},{"name":"device","in":"query","required":false,"schema":{"type":"string","default":"auto","title":"Device"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_and_transcribe_api_asr_upload_and_transcribe_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Upload And Transcribe Api Asr Upload And Transcribe Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/asr/supported-models":{"get":{"tags":["ASR接口"],"summary":"Supported Models","description":"获取支持的模型大小列表\n\n返回:\n- {\"models\": [\"tiny\", \"base\", \"small\", \"medium\", \"large-v2\", \"large-v3\"]}","operationId":"supported_models_api_asr_supported_models_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Supported Models Api Asr Supported Models Get"}}}}}}},"/api/asr/supported-languages":{"get":{"tags":["ASR接口"],"summary":"Supported Languages","description":"获取支持的语言代码列表\n\n返回:\n- {\"languages\": [\"zh\", \"en\", \"ja\", \"ko\", \"fr\", \"de\", ...]}","operationId":"supported_languages_api_asr_supported_languages_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Supported Languages Api Asr Supported Languages Get"}}}}}}},"/api/common/text/clean":{"post":{"tags":["公共功能接口"],"summary":"Clean Text","description":"文本清洗API接口\n\n参数:\n- text: 要清洗的文本内容\n\n返回:\n- 成功: {\"success\": True, \"cleaned_text\": \"清洗后的文本\", \"message\": \"文本清洗成功\"}\n- 失败: {\"success\": False, \"message\": \"错误信息\"}","operationId":"clean_text_api_common_text_clean_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TextCleanRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Clean Text Api Common Text Clean Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/common/text/cut":{"post":{"tags":["公共功能接口"],"summary":"Cut Text","description":"文本切割API接口\n\n参数:\n- text: 要切割的文本内容\n- language: 语言代码，可选值: 'auto', 'zh', 'en', 'fr', 'de', 'ja'（默认: auto）\n\n返回:\n- 成功: {\"success\": True, \"sentences\": [\"句子1\", \"句子2\", ...], \"message\": \"文本切割成功\"}\n- 失败: {\"success\": False, \"message\": \"错误信息\"}","operationId":"cut_text_api_common_text_cut_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TextCutterRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Cut Text Api Common Text Cut Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/common/text/merge":{"post":{"tags":["公共功能接口"],"summary":"Merge Text","description":"文本合并API接口\n\n参数:\n- texts: 要合并的文本列表\n\n返回:\n- 成功: {\"success\": True, \"merged_text\": \"合并后的文本\", \"message\": \"文本合并成功\"}\n- 失败: {\"success\": False, \"message\": \"错误信息\"}","operationId":"merge_text_api_common_text_merge_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TextMergerRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Merge Text Api Common Text Merge Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/common/text/max-split-length":{"get":{"tags":["公共功能接口"],"summary":"Get Max Split Length","description":"获取文本最大分割长度\n\n返回:\n- {\"max_split_length\": 最大分割长度值}","operationId":"get_max_split_length_api_common_text_max_split_length_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Get Max Split Length Api Common Text Max Split Length Get"}}}}}}},"/api/music/list":{"get":{"tags":["音乐接口"],"summary":"Get Music List","operationId":"get_music_list_api_music_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/music/stream/{filename}":{"get":{"tags":["音乐接口"],"summary":"Stream Music","operationId":"stream_music_api_music_stream__filename__get","parameters":[{"name":"filename","in":"path","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/diary/list":{"get":{"tags":["日记接口"],"summary":"Get Diary List","operationId":"get_diary_list_api_diary_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/diary/content/{filename}":{"get":{"tags":["日记接口"],"summary":"Get Diary Content","operationId":"get_diary_content_api_diary_content__filename__get","parameters":[{"name":"filename","in":"path","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/diary/save":{"post":{"tags":["日记接口"],"summary":"Save Diary","operationId":"save_diary_api_diary_save_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DiaryContent"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AudioToTextRequest":{"properties":{"audio_path":{"type":"string","title":"Audio Path"},"model_size":{"type":"string","title":"Model Size","default":"medium"},"language":{"type":"string","title":"Language"},"device":{"type":"string","title":"Device","default":"auto"}},"type":"object","required":["audio_path"],"title":"AudioToTextRequest","description":"音频转文本请求模型（用于本地文件路径）"},"Body_upload_and_transcribe_api_asr_upload_and_transcribe_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_and_transcribe_api_asr_upload_and_transcribe_post"},"ChatRequest":{"properties":{"role":{"type":"string","title":"Role","default":"你是助手"},"message":{"type":"string","title":"Message"},"stream":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Stream","default":false}},"type":"object","required":["message"],"title":"ChatRequest"},"ChatResponse":{"properties":{"response":{"type":"string","title":"Response"}},"type":"object","required":["response"],"title":"ChatResponse"},"DiaryContent":{"properties":{"filename":{"type":"string","title":"Filename"},"content":{"type":"string","title":"Content"},"format":{"type":"string","title":"Format","default":"md"}},"type":"object","required":["filename","content"],"title":"DiaryContent"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"TextCleanRequest":{"properties":{"text":{"type":"string","title":"Text"}},"type":"object","required":["text"],"title":"TextCleanRequest","description":"文本清洗请求模型"},"TextCutterRequest":{"properties":{"text":{"type":"string","title":"Text"},"language":{"type":"string","title":"Language","default":"auto"}},"type":"object","required":["text"],"title":"TextCutterRequest","description":"文本切割请求模型"},"TextMergerRequest":{"properties":{"texts":{"items":{"type":"string"},"type":"array","title":"Texts"}},"type":"object","required":["texts"],"title":"TextMergerRequest","description":"文本合并请求模型"},"TextToAudioRequest":{"properties":{"text":{"type":"string","title":"Text"},"filename":{"type":"string","title":"Filename","default":"output.wav"}},"type":"object","required":["text"],"title":"TextToAudioRequest","description":"文本转语音请求模型"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}